Vehicles_tracking_with_speed_estimation/requirements.txt
torch==1.8.0
PyYAML==6.0
tqdm==4.63.0
opencv-python==4.5.5.64
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.1
seaborn==0.11.2
Pillow==9.0.1
torchvision==0.9.0
requests==2.27.1
yacs==0.1.8
scipy==1.8.0
setuptools==56.0.0
Cython==0.29.28
gdown==4.4.0
easydict==1.9
PyQt5==5.15.4
tensorboard==2.8.0
python-dotenv==0.19.2
boto3==1.21.22
botocore==1.24.22
awscrt==0.13.5
awsiotsdk==1.9.3
